refactor : move directory of menu components
